backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
backports.zoneinfo;python_version<"3.9"
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.9
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-leaflet==0.29.0
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.4.1
importlib-metadata==7.0.1
kombu==5.3.5
Markdown==3.5.2
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
zipp==3.17.0
